Final/src/requirements.txt
efficientnet_pytorch==0.7.1
numpy==1.22.4
opencv_contrib_python==4.5.5.64
Pillow==9.1.1
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
torchvision==0.12.0